Se ha realizado una nueva migración versión 1.2.2, y se empezado a trabajar en el inserción de data en la tabla PagoEmpleado, esto con el fin de facilitar los cálculos de del salario hora de un empleado y así poder llevar el historial...

+++ HorariosConsoleApp/Entities/DetallePagoEmpleado.cs
@@ -5,6 +5,6 @@
         public int DetallePagoEmpleadoId { get; set; }
         public double CantidadHoras { get; set; }
         public string TipoHora { get; set; }
-        public double Porcentaje { get; set; }
+        public decimal Porcentaje { get; set; }
     }
 }
